+++ example/server/cmd/request.go
@@ -31,7 +31,7 @@ var request = &rabbit.Cmd{
 		ctx, cancel := context.WithTimeout(context.Background(), timeout)
 		defer cancel()
 		srv := i.(*node.Server)
-		res, err := srv.Request(ctx, uint16(id), []byte(strings.Join(args, " ")))
+		res, err := srv.Request(ctx, uint32(id), []byte(strings.Join(args, " ")))
 		if err != nil {
 			return err
 		}
